requirements.txt
streamlit
python-dotenv
requests
pandas
fpdf
pydeck
geopy
